Tidy kit/config check/update Split conftest for integration fixtures
