+++ .github/workflows/lint-check.yml
@@ -47,7 +47,7 @@ jobs:
         run: make merge-check-mypy
 
       - name: Run config sync check
-        run: make dev-check-config-sync
+        run: make check-config-sync
 
 # --------------------------------------------------------------------------
 # 2.  Aggregator job — the *single* required status check
