+++ pyproject.toml
@@ -206,7 +206,7 @@ typeCheckingMode = "strict"
 minversion = "9.0"
 addopts = [
     "--import-mode=importlib",
-    "-ra",
+    "-ra",                                                                         # Show all test outcomes (including skips)
     "-m",
     "not (inference or llm or img_gen or extract or needs_output or pipelex_api)",
 ]
