+++ tests/integration/pipelex/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Test fixtures for integration tests."""
