+++ tests/integration/pipelex/fixtures/extract_fixtures.py
@@ -1,0 +1,45 @@
+"""Extract/OCR related test fixtures."""
+
+import pytest
+
+
+@pytest.fixture(
+    params=[
+        "pypdfium2-extract-text",
+        "mistral-ocr",
+    ],
+)
+def extract_handle(request: pytest.FixtureRequest) -> str:
+    assert isinstance(request.param, str)
+    return request.param
+
+
+@pytest.fixture(
+    params=[
+        "mistral-ocr",
+    ],
+)
+def extract_handle_from_image(request: pytest.FixtureRequest) -> str:
+    assert isinstance(request.param, str)
+    return request.param
+
+
+@pytest.fixture(
+    params=[
+        "extract_text_from_visuals",
+        "extract_text_from_pdf",
+    ],
+)
+def extract_choice_for_pdf(request: pytest.FixtureRequest) -> str:
+    assert isinstance(request.param, str)
+    return request.param
+
+
+@pytest.fixture(
+    params=[
+        "extract_text_from_visuals",
+    ],
+)
+def extract_choice_for_image(request: pytest.FixtureRequest) -> str:
+    assert isinstance(request.param, str)
+    return request.param
